src/main/java/com/vilce/handelsbank/models/Watch.java
package com.vilce.handelsbank.models;

public class Watch {
	private String id;
	private String name;
	private Integer price;
	private String discount;
	
	public Watch(String id, String name, Integer price, String discount) {
		super();
		this.id = id;
		this.name = name;
		this.price = price;
		this.discount = discount;
	}
	
	public String getId() {
		return id;
	}
	public void setId(String id) {
		this.id = id;
	}
	public String getName() {
		return name;
	}
	public void setName(String name) {
		this.name = name;
	}
	public Integer getPrice() {
		return price;
	}
	public void setPrice(Integer price) {
		this.price = price;
	}
	public String getDiscount() {
		return discount;
	}
	public void setDiscount(String discount) {
		this.discount = discount;
	}
	
	
}
